Set up CI with Azure Pipelines
Start with an exact copy of the "Python package" template
that has the description "Create and test a Python package
on multiple Python versions" and update it to include:
- [X] Test with Python versions 2.7 and 3.6-3.10 ... the
  template included Python 3.5 but testing on that image
  failed after printing an end-of-life deprecation warning.
- [X] Test on Windows, Linux, and Mac ... very easy, didn't
  require duplicating code
- [X] Publish test results summary to pipeline ... publishes
  named summaries for each job
- [X] Publish code coverage metrics to pipeline ... publishes
  a roll-up
- [X] Add yapf format and flake8 lint check
- [X] Verify test failures fail the pipelines
- [X] Scrub and rename .yml files, pipelines, jobs, stages, and steps

The pipeline structure is a matrix of jobs across different
OSes and Python versions that all run in parallel. All of
these jobs pip install yapf, flake8, pytest, and pytest-cov
and use them to check format, lint, and then execute pytest
and publish test results summaries and code coverage
metrics.

Compare to
- GitHub: https://github.com/Codym48/cmake_saver/pull/1
- GitLab: https://gitlab.com/Codym48/cmake_saver/-/merge_requests/1

Related work items: #1