Autopublish test results via pytest-azurepipelines
This uses pytest 4 and pytest-azurepipelines 0.8.0 to
generate and upload JUnit XML test summary files when
testing with Python 2.7, and pytest 7.1.0 and
pytest-azurepipelines 1.0.3 to generate and upload NUnit
XML test summary files when testing with Python >=3.5
without having to explicitly add publication stages or jobs
or tasks.

This wasn't possible with pytest-azurepipelines 1.0.0-1.0.2
because they required specifying an output file. Version
1.0.3 contains https://github.com/tonybaloney/pytest-azurepipelines/pull/60
which restores the generation of a test summary output file
without requiring any additional runtime arguments, a
closer match to the 0.8.0 behavior.

Related work items: #1