List bad files by default, only fix if given --fix
Add requirements, imlementation, and unit tests to verify that
calling cmake_saver[.py] only fixes files in place if given the
--fix (-f) input.

Without this argument, cmake_saver[.py] will list all CMake files
in the input directory that contain trailing whitespace characters
and return the number of files in that list as an error code. This
supports a newly documented use case of calling cmake_saver[.py] in
pipelines to block introduction of trailing whitespace.

I considered an alternate approach that required one input of
either --fix (-f) or --list (-l) to force the user to be explicit,
but this didn't seem necessary.